I created a forumAPI (read description)
I allowed for the creation of users and threads. On threads, there are users attached. Every thread and user has a unique ID. There are built-in exceptions to users or threads not existing. If a thread is posted with a user that is not currently in the dictionary, then they will be added.